✨ Add User Permission utilities
