+++ types/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+  isLoggedIn: boolean;
+  login: string;
+  avatarUrl?: string;
+  admin?: true;
+};
